Some first steps in the direction of a plugin system.
Let the backends be the only visible things outside of Coherence!

We should now be able to add and activate a plugin an startup and, in the next steps, later via the WebUI too, and retrieve from somewhere the configuration for each plugin.

And this foundation should be extendable for .egg plugins too.