Cleaned up the backend upnp_... functions, moved the universal ones into the service.
This addresses #36, but we need for this to work properly some additional mapping supplied by the backend to tell which of the Windows Media Connect 'directories' the backend actually supports.

Added a method to browse the content of the Flickr backend.
This should evolve in some future into something that allows browser access to any of the non-filesystem backends.