Fix severe bug: SSDP server holds all known svcs and callbacks globally.
This was as the attributes `known` and `_callbacks` have been class-
attributes. Thus every SSDP server instance within one program had
the *same* services and callbacks registered. While this seams not
to have shown up as an error yet. It is at least bad design if not
severely buggy.

Conflicts:
	coherence/upnp/core/ssdp.py