Fix FSStore
In the current project state, only the dlna/upnp client is working as expected, the FSStore feature fails due to migration to python3. The big change related to the string management introduced in python 3 (unicode is the default encoding string, instead of ascii) affects the interaction with twisted dependency, because of this python 3 change, the twisted 3 package migrated some of the variables to be bytes (ascii)...so...we are enforced to apply some changes in order to make the FSStore feature to work again. Changing those string to be ascii encoded solves the issue, but this affects a lot of files as well some of the tests...so...here are some of the most important stuff that should be mentioned:

  - service variable: control_url (now is ascii)
  - service variable: event_sub_url (now is ascii)
  - service variable: presentation_url (now is ascii)
  - service variable: url_base (now is ascii)
  - putChild function first argument is now an ascii encoded string
  - all the header's request are now byte strings
  - fix some typo errors
  - add some log messages (some of them disabled because they are not usually needed)
  - request.received_headers replaced by request.responseHeaders
  - removes store SimpleLight from docs, because this store was into the original Coherence project and was removed at some point...when the project was migrated to Cohen
  - All the tests has been updated to match the new circumstances, including the dbus test which has been upgraded to use the GIReactor (instead the old Glib2Reactor). The test_dbus has been disabled because causes the utils.getPage test to fail, despite that when running this test alone seems to work without problems (this issue should be reviewed at some point)

 With this commit, we restore the basic functionality of the project: create a dlna/upnp client and create a dlna/upnp server.