Add external dependencies non migrated to python 3 and fix related imports
Some of the plugins rely on some twisted packages non migrated to python 3: axiom and epsilon. This commit adds the mentioned dependencies partially migrated to python 3, partially because those should be tested with the affected plugins and probably modify some lines, despite that all the code has been reviewed after migrated to python 3, those aren't fully tested. The main purpose of adding this files are to making work the Cohen 3 project as fast as possible, without depending on external dependencies non migrated to python 3. Be aware that all the imports of this added packages are now relative to Cohen3 and they are not compatible with python 2 without some work. Those added packages...maybe should be replaced for the original packages when/if updated to be compatible with python 3.