Enhance setup with extras_require
In order make easier to install the package, depending of the use of the framework, the setup.py file has been modified. As well, with this commit, update the README.rst file with new instructions and adds the current status of the project. Also updates the travis.yml file to match the new setup.py file which adds more travis tests skipped before because of the lack of some dependencies, and sets nosetests as default travis builds (as it was in the initial project).