Fix pylint E1101 for youtube modules
Despite the fixed pylint errors, this backend will not work for now...but at least will allow to pass the pylint errors. The problems of this modules are beyond the actual source code because google released a new api which is not implemented here and the old one has been shutdown.

References: #11