Add utilities and improve documentation.
Add string encode/decode utilities to make easier to encode/decode strings sent/received from/to server. This also will simplify the source code and make easier to read it.

Also, some of the documentation of the affected file is not compliance accordingly our guidelines. Here we take care of that and add some more documentation for undocumented modules, classes and methods.