Whole new design for web server html visualization
Add ability to to show the server content styled (with css stylesheet). Adapt the render methods to make use of this stylesheet and also adds the ability to show the project's icon, in order to improve the whole visualization when making use of the html visualization for a webserver. The added stylesheet is dark themed.