Reintroduces WebUI
The WebUI was previously removed. Here we restore the web ui capabilities with a whole new approach. This will remove a project dependency (twisted's Nevow) which isn't compatible with python 3 and we will do all the communication between python and the browser relying on web sockets and using the help of the module autobahn and twisted. Should be mentioned that the connection with the reactor is done via twisted.internet.endpoints instead of reactor.listenTCP (Low Level Api) because it seems that the endpoint approach it's more maintainable and versatile than the old approach, and moreover, it's the recommended way these days.

References: 3ffac64