Fix Request for coherence.upnp.core.utils
In our process method of Request we should not always return bytes, because some times our processed result it will not be an instance of a string or bytes string and we want that the render function resolves the situation. This happens due to recent change on to_bytes function.

References: 2a0d92a