Add parse_with_lxml and fix encoding for parse_xml
The added function "parse_with_lxml" allow us to fix the parsing of the soap messages with encoding declared. The parse_xml function has an encode call that will lead to errors if successfully applied because the returned result from the encode will not by of type bytes, which we need for the following operations.