Specify the eventdispatcher version in setup.py.
Unfortunately the eventdispatcher project uses an odd version number
sequence, for example `1.81` is older than `1.9.4`. This causes issues
with determining the latest version and can result in non-working
installs.

This version was set to match the version used elsewhere in the project
for testing, although it's possible a wider range would also be valid.